+++ client/src/Components/LoginPage/index.jsx
@@ -1,15 +1,35 @@
 import React, { useState } from "react";
 import { TextField, InputAdornment, Button, Typography } from "@mui/material";
-
+import { useFormik } from "formik";
+import * as Yup from "yup";
 import { Visibility, VisibilityOff, Email } from "@mui/icons-material";
 import PropTypes from "prop-types";
 
-export default function Login({ loginForm }) {
+const loginSchema = Yup.object().shape({
+  email: Yup.string()
+    .email(" البريد الالكتروني غير صحيح")
+    .required("حقل البريد الالكتروني مطلوب"),
+  password: Yup.string()
+    .min(8, "كلمة المرور على الاقل 8 احرف")
+    .required("حقل كلمة المرور مطلوب"),
+});
+
+export default function Login({ onFinish }) {
   const [showPassword, setShowPassword] = useState(false);
 
   const handleClickShowPassword = () => {
     setShowPassword(!showPassword);
   };
+  const loginForm = useFormik({
+    initialValues: {
+      email: "",
+      password: "",
+    },
+    validationSchema: loginSchema,
+    onSubmit: (values) => {
+      onFinish(values);
+    },
+  });
   return (
     <form className="form__container" onSubmit={loginForm.handleSubmit}>
       <Typography className="center" variant="h3" component="h4" sx={{ mb: 2 }}>
@@ -66,5 +86,5 @@ export default function Login({ loginForm }) {
   );
 }
 Login.propTypes = {
-  loginForm: PropTypes.instanceOf(Object).isRequired,
+  onFinish: PropTypes.func.isRequired,
 };
