+++ client/src/Pages/LoginGym.jsx
@@ -2,18 +2,9 @@ import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { Typography, Snackbar, Alert } from "@mui/material";
 import axios from "axios";
-import { useFormik } from "formik";
-import * as Yup from "yup";
+
 import { Login } from "../Components";
 
-const loginSchema = Yup.object().shape({
-  email: Yup.string()
-    .email(" البريد الالكتروني غير صحيح")
-    .required("حقل البريد الالكتروني مطلوب"),
-  password: Yup.string()
-    .min(8, "كلمة المرور على الاقل 8 احرف")
-    .required("حقل كلمة المرور مطلوب"),
-});
 export default function LoginGym() {
   const [state, setState] = useState({
     message: "",
@@ -41,21 +32,10 @@ export default function LoginGym() {
     }
   };
 
-  const loginForm = useFormik({
-    initialValues: {
-      email: "",
-      password: "",
-    },
-    validationSchema: loginSchema,
-    onSubmit: (values) => {
-      onFinish(values);
-    },
-  });
-
   return (
     <div className="container">
       <div className="sub__container">
-        <Login loginForm={loginForm} />
+        <Login onFinish={onFinish} />
         <Typography className="center" component="h4" sx={{ mt: 2 }}>
           ليس لديك حساب ؟&ensp;<Link to="/gym/register"> إنشاء حساب </Link>
         </Typography>
